Integrate Sinatra test, and fix up the Ruby package.
