+++ .bumpversion.cfg
@@ -28,3 +28,5 @@ search = "package datawire_mdk_util {current_version};"
 [bumpversion:file:javascript/datawire_mdk_express/package.json]
 search = '"version": "{current_version}",'
 
+[bumpversion:file:ruby/rack-mdk/rack-mdk.gemspec]
+search = "s.version     = '{current_version}'"
