+++ ruby/rack-mdk/lib/rack-mdk.rb
@@ -1,0 +1,1 @@
+require 'rack/mdk'
