+++ ruby/rack-mdk/rack-mdk.gemspec
@@ -3,7 +3,7 @@ Gem::Specification.new do |s|
   s.version     = '2.0.13'
   s.summary     = "Datawire MDK integration Rack middleware."
   s.authors     = ["Datawire Inc."]
-  s.files       = ["lib/rackmdk.rb"]
+  s.files       = ["lib/rack-mdk.rb", "lib/rack/mdk.rb"]
   s.homepage    = 'https://github.com/datawire/mdk'
   s.license     = 'Apache-2.0'
 end
